modules/apps/commerce/commerce-currency-api/src/main/java/com/liferay/commerce/currency/util/CommercePriceFormatter.java
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

package com.liferay.commerce.currency.util;

import com.liferay.commerce.currency.model.CommerceCurrency;
import com.liferay.portal.kernel.exception.PortalException;

import java.math.BigDecimal;

import java.util.Locale;

import javax.portlet.ActionRequest;

/**
 * @author Marco Leo
 * @author Alessio Antonio Rendina
 */
public interface CommercePriceFormatter {

	public String format(BigDecimal price, Locale locale)
		throws PortalException;

	public String format(
			CommerceCurrency commerceCurrency, BigDecimal price, Locale locale)
		throws PortalException;

	public String formatAsRelative(
		CommerceCurrency commerceCurrency, BigDecimal price, Locale locale);

	public BigDecimal parse(ActionRequest actionRequest, String param)
		throws Exception;

	public String parse(String price, Locale locale) throws Exception;

}